pyproject.toml
[project]
name = "dfc-task-presence"
version = "0.1.0"
description = "Exploratory project to improve classification of task and rest time points using dFC data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "appnope==0.1.4 ; sys_platform == 'darwin'",
    "asttokens==3.0.0",
    "async-timeout==5.0.1 ; python_full_version < '3.11'",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "certifi==2025.8.3",
    "cffi==1.17.1 ; implementation_name == 'pypy'",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "comm==0.2.3",
    "contourpy==1.3.2 ; python_full_version < '3.11'",
    "contourpy==1.3.3 ; python_full_version >= '3.11'",
    "contrastive==1.2.0",
    "cycler==0.12.1",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "exceptiongroup==1.3.0 ; python_full_version < '3.11'",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.13.1",
    "fonttools==4.59.0",
    "frozenlist==1.7.0",
    "fsspec==2024.6.1",
    "hf-xet==1.1.7 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'",
    "huggingface-hub==0.34.3",
    "idna==3.10",
    "importlib>=1.0.4",
    "ipykernel==6.30.1",
    "ipython==8.37.0 ; python_full_version < '3.11'",
    "ipython==9.4.0 ; python_full_version >= '3.11'",
    "ipython-pygments-lexers==1.1.1 ; python_full_version >= '3.11'",
    "jedi==0.19.2",
    "jinja2==3.1.4",
    "joblib==1.5.1",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyterlab-pygments==0.3.0",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.5",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "mpmath==1.3.0",
    "multidict==6.6.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.3",
    "numpy==2.1.2",
    "nvidia-cublas-cu12==12.6.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12==12.6.80 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12==12.6.77 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12==12.6.77 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.5.1.17 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.3.0.4 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cufile-cu12==1.11.1.6 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-curand-cu12==10.3.7.77 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusolver-cu12==11.7.1.2 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusparse-cu12==12.5.4.2 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusparselt-cu12==0.6.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nccl-cu12==2.26.2 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12==12.6.85 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvtx-cu12==12.6.77 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "packaging==25.0",
    "pandas==2.3.1",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pillow==11.0.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "psutil==7.0.0",
    "ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pure-eval==0.2.3",
    "pycparser==2.22 ; implementation_name == 'pypy'",
    "pygments==2.19.2",
    "pylance>=0.33.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pywin32==311 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rpds-py==0.27.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.1",
    "scipy==1.15.3 ; python_full_version < '3.11'",
    "scipy==1.16.1 ; python_full_version >= '3.11'",
    "seaborn==0.13.2",
    "setuptools==70.2.0 ; (python_full_version >= '3.12' and platform_machine != 'x86_64') or (python_full_version >= '3.12' and sys_platform != 'linux') or (platform_machine == 'x86_64' and sys_platform == 'linux')",
    "six==1.17.0",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "sympy==1.13.3",
    "tensorly==0.9.0",
    "threadpoolctl==3.6.0",
    "timm==1.0.19",
    "tinycss2==1.4.0",
    "torch==2.7.1",
    "torch-geometric==2.6.1",
    "torchaudio==2.7.1",
    "torchvision==0.22.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.3.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "typing-extensions==4.12.2",
    "tzdata==2025.2",
    "umap==0.1.1",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "yarl==1.20.1",
]
